feat: Add URL param filter option.
This allows query parameters to be set, so that we can craft links to
the backpack that show only badges that belong to certain badge class
identifiers.

e.g. backpack?filter.badgeclass_id=vwDwRDTbR_yeTV-r6m1lnw&filter.badgeclass_id=CiM7MSFASeGIrywCQlyjhg

If a filter is active, we show this to the user and allow them to
"remove the filter", ie. link to the /backpack.

There is no UI to manage the filters. by design.