fix: Also update the filtered badges when already on /backpack
When navigating from e.g. backpack?filter.badgeclass_id=11BR-001-01&filter.badgeclass_id=11BR-001-02
to backpack, onMount is not re-evaluated.
So we must manually update some variables when clicking the link.

For that, we've refactored the loading to be easier to follow and more
modular. We needed this because we don't want to trigger a full rebuild.
Now we can change only the variables that must be changed.