refactor: Clean and fix redirect handling logic
Refactor the redirect handling logic so that we can debug it.
By moving some parts to a central place and re-use some logic,
we can determine that the Login router was re-setting the $redirectPath
to "/login" on hitting the login page.

We've now fixed this. And on return are redirected to the page that we
got the access denied on.

remaining issue: If a user hits a path only accessible to **students**,
logs in with a **staff** role, the redirect will land them on a page
that they have no access to, which then triggers the login logic. Same
with staff path and login as user.

This can, and most likely will, put the user in an endless redirect
loop! This issue existed in the previous logic. But might not have
surfaced because the redirects didn't work properly. The chances of this
triggering is small though: it will only exist with users that both have
a backpack as student and a staff account as staff, and then only when
they mix up logins.