feat: Simplify the allowed markdown.
We cannot remove hardcoded items from the WYSIWYG editor toolbar, so we
have to keep "code", "codeblock" and "link", despite not supporting them
in the actual credential json.

This allows us to remove the GFM that provided a.o. tables. Tables are
not allowed in our current setup.

And I think tables weren't supported in the actual badges before either,
as the backend does not allow any of the table, tr, td etc tags.